Make profile module

+++ app/Http/Controllers/Admin/PlanController.php
@@ -70,11 +70,11 @@ class PlanController extends Controller
         if (!$plan)
             return redirect()->back();
 
-        // if ($plan->details->count() > 0) {
-        //     return redirect()
-        //         ->back()
-        //         ->with('error', 'Existem detahes vinculados a esse plano, portanto não pode deletar');
-        // }
+        if ($plan->details->count() > 0) {
+            return redirect()
+                ->back()
+                ->with('error', 'Existem detahes vinculados a esse plano, portanto não pode deletar');
+        }
 
         $plan->delete();
